requirement.txt
Flask==1.1.1
keras==2.3.0
scikit-learn==0.23.2
flask_restfull==0.3.8